Extract attribute related code into own module
